reorganize structures test
